fix: check if ssh is available for ipv4 that is found
When removing package linux-firmware with updated latest kernel package,
the `virsh domifaddr` shows 2 interfaces with two IPs and one MAC - the
one with MAC has wrong IP - and taht's the one we try to wait for. This
hangs forever. However the second works and also by `arp in <MAC>` we
get the working IP. But in the python script, blindly trying with
timeout the ssh was the easiest and fastest to implement.